+++ src/pvaccess/PvObject.h
@@ -47,6 +47,9 @@ public:
     friend std::ostream& operator<<(std::ostream& out, const PvObject& pvObject);
     friend epics::pvData::PVStructurePtr& operator<<(epics::pvData::PVStructurePtr& pvStructurePtr, const PvObject& pvObject);
 
+    // Copy
+    PvObject copy();
+
     // Dictionary methods
     bool has_key(const std::string& fieldPath) const;
     boost::python::list items() const;